+++ iogt/urls.py
@@ -7,6 +7,8 @@ from wagtail.admin import urls as wagtailadmin_urls
 from wagtail.core import urls as wagtail_urls
 from wagtail.documents import urls as wagtaildocs_urls
 
+from iogt.views import ExternalLink
+
 urlpatterns = [
     path("django-admin/", admin.site.urls),
     path("admin/", include(wagtailadmin_urls)),
@@ -14,6 +16,7 @@ urlpatterns = [
     path("search/", search_views.search, name="search"),
     path("users/", include(users_urls), name="users_urls"),
     path("accounts/", include("allauth.urls"), name="allauth-urls"),
+    path("external-link/", ExternalLink.as_view(), name="external_link"),
 ]